server/pyproject.toml
[project]
name = "vittam-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "langchain>=1.1.3",
    "langchain-mongodb>=0.9.0",
    "langchain-openai>=1.1.3",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
